added benchmark for doubles to saxpy demos

+++ demo/saxpy/nt2/saxpy_nt2.cpp
@@ -65,7 +65,7 @@ template<typename T> struct axpy_nt2
     nt2::table<T> X, Y;
 };
 
-NT2_REGISTER_BENCHMARK_TPL( axpy_nt2, (float) )
+NT2_REGISTER_BENCHMARK_TPL( axpy_nt2, (float)(double) )
 {
   std::size_t size_min = args("size_min", 16);
   std::size_t size_max = args("size_max", 4096*8);
